pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "starmaker-toolkit"
version = "0.1.0"
description = "Universal OSS promotion toolkit — generate post drafts, audit repos, find awesome-lists, and enhance READMEs"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    {name = "Muminur", email = "muminur@users.noreply.github.com"},
]
keywords = ["github", "open-source", "promotion", "cli", "stars"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "requests>=2.28",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
]

[project.scripts]
starmaker = "starmaker.cli:main"

[project.urls]
Homepage = "https://github.com/Muminur/starmaker-toolkit"
Repository = "https://github.com/Muminur/starmaker-toolkit"
Issues = "https://github.com/Muminur/starmaker-toolkit/issues"

[tool.setuptools.packages.find]
include = ["starmaker*"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.pytest.ini_options]
testpaths = ["tests"]
